requirements.txt
# Async & Networking (NON-NEGOTIABLE)
aiohttp>=3.9.0
websockets>=12.0

# Data Processing (Speed Matters)
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
numba>=0.58.0

# Math & Statistics
scipy>=1.11.0
statsmodels>=0.14.0
pykalman>=0.9.5
filterpy>=1.4.5

# Machine Learning
scikit-learn>=1.3.0

# Visualization (The "Wow" Factor)
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
hypothesis>=6.92.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.1

# HTTP Requests (fallback)
requests>=2.31.0
